Add special instructions for word processors
To warn users about the disappearing Ra symbol, which
is bound to the apostrophy `'` character. If the word
processor converts it to a right single quote `’` the
visual character can mysteriously disappear from the
text.

In addition to this warning, we should consider making
an augmented version of the font that has Ra for the
right single quote character as well, removing the
issue for new users.