Allow wide compositions to scroll independently
Previously, the width of compositions would add an
extra width to the entire page, which would make
scrolling on mobile awkward. This would only happen
in the Bhatkhande pages, which had their wide sample
composition in a table. In Swarlipi, where the sample
is in a pre block, it would scroll horizontally within
the block, without making the entire page wider.

By wrapping the table in a div with the same CSS as the
pre block, we allow for wide compositions that can be
scrolled independently. Furthermore, this allows us to
restore the font size of the compositions to be more
in line with the rest of the page.