Validate via FontForge
Remove non-integral values for shapes, add missing
extremities, ensure there are no self-intersecting
shapes, and generally validate the font.

This is done mainly by the auto-fix option in
FontForge, with some hand-tweaking done for `n`
and `N` to ensure the lack of self-intersection,
by making sure the chhat-wali-lakir of the Na does
not intersect with that of the ee-ki-matra.