Update FontForge to 2023-01-01
We are now getting contributions that target more modern file
formats, e.g. SFD v3.2 in #59, and need to support more recent
versions of FontForge than the current version which is more
than 8 years old.

I was unable to find a pre-built Docker image on Dockerhub for
fontforge, so ended up making a simple Dockerfile based on
Ubuntu:22.04 and adding it to docker-compose.yml myself. This
has the additional advantage of being an ARM image on Apple
Silicon macOS, which is great for performance.