Refactor build script
Rewrite convert script in Python, which is a lot
better than the fontforge script.

Add a Bhatkhande build script, and use it in the
general build process.

Amend the build script to take the Bhatkhande
base font, then layer in a language, then use it
to generate .UFO and .TTF files. The intermediate
.SFD files are deleted, so as not to cause any
confusion.