Add interim version of Bangla
This uses contributions from @jain-sanjay and regularizes them.
It also pulls glyphs from AkaashNormal, the source font he used
for the Bangla glyphs. Unfortunately, the glyphs are in different
scales, and don't line up correctly.

The next step is to re-source all the glyphs directly from
AkaashNormal so that they are even.